+++ lib/common/navigator/bottom_navigation_bar_view_model.dart
@@ -1,5 +1,6 @@
 import 'package:cogo/common/enums/role.dart';
 import 'package:cogo/features/home/home_view_model.dart';
+import 'package:cogo/features/mypage/profile_management/mentor_introduction_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
@@ -41,15 +42,32 @@ class BottomNavigationViewModel extends ChangeNotifier {
     }
   }
 
-  // 다이얼로그를 띄우는 함수
+  /// 다이얼로그를 띄우는 함수
   void _showMentorProfileDialog(BuildContext context) {
     showDialog(
       context: context,
       builder: (BuildContext context) {
-        return const InfoDialog(
+        return BasicDialog(
           title: "멘토 활동을 시작하려면\n프로필 작성을 완료해주세요",
           subtitle: '입력하신 정보는 하단의 MY에서 수정이 가능해요',
           imagePath: 'assets/icons/3d_img/heart.png',
+          buttonText: '멘토 프로필 작성하기',
+          onPressed: () => Navigator.of(context).push(_createRoute()),
+        );
+      },
+    );
+  }
+
+  Route _createRoute() {
+    return PageRouteBuilder(
+      pageBuilder: (context, animation, secondaryAnimation) =>
+          const MentorIntroductionScreen(),
+      transitionsBuilder: (context, animation, secondaryAnimation, child) {
+        var tween = Tween(begin: const Offset(0.0, 1.0), end: Offset.zero)
+            .chain(CurveTween(curve: Curves.ease));
+        return SlideTransition(
+          position: animation.drive(tween),
+          child: child,
         );
       },
     );
